restructure spring-boot classes
