+++ upday/webservice/src/main/java/com/upday/service/Application.java
@@ -1,4 +1,4 @@
-package com.upday.service.web;
+package com.upday.service;
 
 import java.util.Arrays;
 
@@ -7,8 +7,10 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.builder.SpringApplicationBuilder;
 import org.springframework.boot.context.web.SpringBootServletInitializer;
 import org.springframework.context.ApplicationContext;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
 @SpringBootApplication
+@EnableJpaRepositories
 public class Application extends SpringBootServletInitializer {
     
     public static void main(String[] args) {
