+++ upday/webservice/src/main/java/com/upday/service/web/WebServiceInitializer.java
@@ -1,29 +1,21 @@
 package com.upday.service.web;
 
-import com.upday.service.web.WebConfig;
+import com.upday.service.web.config.WebConfig;
 import javax.servlet.ServletContext;
 import javax.servlet.ServletRegistration;
  
 import org.springframework.web.WebApplicationInitializer;
-import org.springframework.web.context.ContextLoaderListener;
 import org.springframework.web.context.support.AnnotationConfigWebApplicationContext;
-import org.springframework.web.filter.DelegatingFilterProxy;
 import org.springframework.web.servlet.DispatcherServlet;
  
 /**
- *
+ * //TODO
  * @author jschulz
  */
 public class WebServiceInitializer implements WebApplicationInitializer {
  
     @Override
     public void onStartup(ServletContext container) {
-        // Create the 'root' Spring application context
-//        AnnotationConfigWebApplicationContext rootContext = new AnnotationConfigWebApplicationContext();
-//        rootContext.register(ServiceConfig.class, JPAConfig.class, SecurityConfig.class);
- 
-        // Manage the lifecycle of the root application context
-//        container.addListener(new ContextLoaderListener(rootContext));
  
         // Create the dispatcher servlet's Spring application context
         AnnotationConfigWebApplicationContext dispatcherServlet = new AnnotationConfigWebApplicationContext();
