+++ upday/webservice/src/test/java/com/upday/service/web/HelloControllerIT.java
@@ -1,5 +1,6 @@
 package com.upday.service.web;
 
+import com.upday.service.Application;
 import static org.hamcrest.Matchers.equalTo;
 import static org.junit.Assert.assertThat;
